+++ tests/test_yaralyze.py
@@ -20,7 +20,7 @@ from tests.yara.test_yara_rule_builder import HEX_STRING
 def test_help_option():
     help_text = _run_with_args('-h')
     assert 'maximize-width' in help_text
-    _assert_line_count_within_range(111, help_text)
+    _assert_line_count_within_range(118, help_text)
 
 
 def test_no_rule_args(il_tulipano_path):